Added gyroscopic movement
Tilt controls work for movement.
Implemented variable for sensitivity.

For some reason phones don't register gyroscope as an input, so the phone would sleep timeout in the middle of play, so I added a line of code to prevent it. Later we can specify that the game can time out in menus, but not in-game.